HttpProviders: create separate JavaHttpProvider, move ApacheHttpClientProvider to here-apache-httpprovider module.
Separate out IT tests from Unit tests.  Test coverage boosts are needed.
