+++ here-apache-httpprovider/pom.xml
@@ -1,24 +1,43 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <name>HERE OAuth Client Example</name>
-    <artifactId>here-oauth-client-example</artifactId>
+    <name>HERE Apache HttpClient Provider</name>
+    <artifactId>here-apache-httpprovider</artifactId>
     <packaging>jar</packaging>
-    <description>here-oauth-client-example is for examples which use here-oauth-client</description>
+    <description>here-apache-httpprovider using Apache HttpClient library for transport</description>
 
     <parent>
         <groupId>com.here.account</groupId>
         <artifactId>here-aaa-sdk</artifactId>
         <version>0.4.7-SNAPSHOT</version>
-        <relativePath>../../pom.xml</relativePath>
+        <relativePath>../pom.xml</relativePath>
     </parent>
 
     <build>
+        <!-- include the filtered resources in the jar -->
+        <resources>
+            <resource>
+                <filtering>true</filtering>
+                <directory>src/main/resources</directory>
+            </resource>
+        </resources>
+
         <!-- activate these plugins in this build -->
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
+            </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-jar-plugin</artifactId>
+                <executions>
+                    <execution>
+                        <goals>
+                            <goal>test-jar</goal>
+                        </goals>
+                    </execution>
+                </executions>
             </plugin>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
@@ -56,26 +75,6 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-failsafe-plugin</artifactId>
             </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-shade-plugin</artifactId>
-                <version>2.4.3</version>
-                <executions>
-                  <execution>
-                    <phase>package</phase>
-                    <goals>
-                      <goal>shade</goal>
-                    </goals>
-                    <configuration>
-                      <transformers>
-                        <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                          <mainClass>com.here.account.oauth2.tutorial.GetHereClientCredentialsAccessTokenTutorial</mainClass>
-                        </transformer>
-                      </transformers>
-                    </configuration>
-                  </execution>
-                </executions>
-            </plugin>
         </plugins>
     </build>
 
@@ -88,16 +87,10 @@
         </dependency>
         <dependency>
             <groupId>com.here.account</groupId>
-            <artifactId>here-apache-httpprovider</artifactId>
+            <artifactId>here-oauth-client</artifactId>
             <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>commons-codec</groupId>
-            <artifactId>commons-codec</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.fasterxml.jackson.core</groupId>
-            <artifactId>jackson-databind</artifactId>
+            <type>test-jar</type>
+            <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>org.apache.httpcomponents</groupId>
