+++ here-apache-httpprovider/src/main/java/com/here/account/http/apache/ApacheHttpClientProvider.java
