+++ here-apache-httpprovider/src/test/java/com/here/account/http/apache/ApacheHttpClientProviderTest.java
