+++ here-oauth-client-dist/pom.xml
@@ -68,6 +68,17 @@
         </dependency>
         <dependency>
             <groupId>com.here.account</groupId>
+            <artifactId>here-apache-httpprovider</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.here.account</groupId>
+            <artifactId>here-apache-httpprovider</artifactId>
+            <classifier>javadoc</classifier>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.here.account</groupId>
             <artifactId>here-oauth-client-example</artifactId>
             <version>${project.version}</version>
         </dependency>
