+++ here-oauth-client/src/main/java/com/here/account/http/java/JavaHttpProvider.java
