+++ here-oauth-client/src/main/java/com/here/account/util/JsonSerializer.java
@@ -95,7 +95,8 @@ public class JsonSerializer {
      * 
      * @param object the object to serialize into a JSON string.
      * @return the JSON string representation of the object.
-     * @throws JsonProcessingException 
+     * @throws JsonProcessingException if there's trouble serializing object 
+     *      to a JSON string.
      */
     public static String objectToJson(Object object) throws JsonProcessingException {
         return objectMapper.writeValueAsString(object);
