+++ here-oauth-client/src/test/java/com/here/account/oauth2/AbstractHereAccountProviderTezt.java
@@ -69,6 +69,4 @@ public abstract class AbstractHereAccountProviderTezt extends AbstractCredential
         assertTrue("accessToken was null or zero-length", null != accessToken && accessToken.length() > 0);
     }
 
-    protected abstract HttpProvider getHttpProvider();
-
 }
